test: consolidate story-related tests by replacing outdated files with new implementations
