+++ test/app/features/stories/providers/story_controllers_test.dart
@@ -2,10 +2,15 @@
 
 import 'package:flutter_test/flutter_test.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:ion/app/features/feed/stories/data/models/story.c.dart';
+import 'package:ion/app/features/auth/providers/auth_provider.c.dart';
 import 'package:ion/app/features/feed/stories/providers/stories_provider.c.dart';
 import 'package:ion/app/features/feed/stories/providers/story_viewing_provider.c.dart';
+import 'package:ion/app/features/feed/stories/providers/viewed_stories_provider.c.dart';
+import 'package:ion/app/services/storage/local_storage.c.dart';
+import 'package:ion/app/services/storage/user_preferences_service.c.dart';
+import 'package:mocktail/mocktail.dart';
 
+import '../../../../mocks.dart';
 import '../helpers/story_test_models.dart';
 import '../helpers/story_test_utils.dart';
 
@@ -15,15 +20,8 @@ void main() {
   const alice = 'alice';
   const bob = 'bob';
 
-  final aliceUserStories = UserStories(
-    pubkey: alice,
-    stories: [buildPost('a1'), buildPost('a2')],
-  );
-
-  final bobUserStories = UserStories(
-    pubkey: bob,
-    stories: [buildPost('b1')],
-  );
+  final aliceUserStories = buildUserStories(alice, ['a1', 'a2']);
+  final bobUserStories = buildUserStories(bob, ['b1']);
 
   ProviderContainer createContainer() => createStoriesContainer(
         overrides: [
@@ -62,7 +60,7 @@ void main() {
 
       final state = container.read(storyViewingControllerProvider(alice));
 
-      expect(state.currentUserIndex, 1); // Bob
+      expect(state.currentUserIndex, 1);
       expect(state.currentStoryIndex, 0);
     });
 
@@ -99,9 +97,7 @@ void main() {
       'with the current user (CubePageView onPageChanged)',
       () {
         final container = createContainer();
-        final notifier = container.read(storyViewingControllerProvider(alice).notifier)
-          // Alice/a1 → Alice/a2  (user = 0, story = 1)
-          ..advance();
+        final notifier = container.read(storyViewingControllerProvider(alice).notifier)..advance();
 
         var state = container.read(storyViewingControllerProvider(alice));
         expect(state.currentUserIndex, 0);
@@ -114,5 +110,74 @@ void main() {
         expect(state.currentStoryIndex, 1);
       },
     );
+  });
+
+  group('ViewedStoriesController', () {
+    const identity = 'user1';
+    const keyPrefix = 'user_$identity:';
+    const prefKey = '${keyPrefix}viewedStories';
+
+    late MockLocalStorage mockStorage;
+
+    setUp(() {
+      mockStorage = MockLocalStorage();
+      when(() => mockStorage.setStringList(any(), any<List<String>>()))
+          .thenAnswer((_) async => true);
+    });
+
+    ProviderContainer buildContainer({List<String>? initialIds}) {
+      when(() => mockStorage.getStringList(prefKey)).thenReturn(initialIds);
+
+      return createStoriesContainer(
+        overrides: [
+          currentIdentityKeyNameSelectorProvider.overrideWith((_) => identity),
+          localStorageProvider.overrideWithValue(mockStorage),
+          userPreferencesServiceProvider(identityKeyName: identity).overrideWith(
+            (_) => UserPreferencesService(identity, mockStorage),
+          ),
+        ],
+      );
+    }
+
+    test('build() reads saved Set<String> from preferences', () {
+      final container = buildContainer(initialIds: ['s1', 's2']);
+
+      final state = container.read(viewedStoriesControllerProvider);
+
+      expect(state, equals({'s1', 's2'}));
+      verify(() => mockStorage.getStringList(prefKey)).called(1);
+    });
+
+    test('markStoryAsViewed() persists only on first insert', () async {
+      final container = buildContainer(initialIds: []);
+
+      final notifier = container.read(viewedStoriesControllerProvider.notifier);
+
+      await notifier.markStoryAsViewed('storyX');
+      expect(
+        container.read(viewedStoriesControllerProvider),
+        contains('storyX'),
+      );
+      verify(() => mockStorage.setStringList(prefKey, ['storyX'])).called(1);
+
+      await notifier.markStoryAsViewed('storyX');
+      verifyNever(
+        () => mockStorage.setStringList(prefKey, ['storyX', 'storyX']),
+      );
+    });
+
+    test('syncAvailableStories() removes stale IDs and persists once', () async {
+      final container = buildContainer(initialIds: ['a', 'b', 'c']);
+
+      final notifier = container.read(viewedStoriesControllerProvider.notifier);
+
+      await notifier.syncAvailableStories(['a', 'c']);
+      verify(() => mockStorage.setStringList(prefKey, ['a', 'c'])).called(1);
+
+      clearInteractions(mockStorage);
+
+      await notifier.syncAvailableStories(['a', 'c']);
+      verifyNever(() => mockStorage.setStringList(any(), any()));
+    });
   });
 }
